refactor(chat): establish components/ and inventory/ layout
Cleaner file layout for the Preact migration:

- inventory/ — the inventory feature's code, dropping the inventory- prefix:
  inventory.js (was inventory-component.js), dnd.js, tree-source.js,
  drop-menu.js, item-actions.js.
- components/ — reusable, app-wide Preact components. Empty for now (no piece
  has a second caller yet); components/README.md documents the one common
  component format both directories follow.
- setup-preact-container.js stays at root as the app barrel.

Pure move: imports updated across chat.js, the moved files, the AVA component
test, and the browser-test probes; tsconfig includes the new dirs. No
behavior change. The design doc gains a File layout section.

Verified: eslint (0 errors), prettier, tsc neutral (chat.js 30 and
inventory/inventory.js 9 errors are pre-existing, unchanged); AVA
inventory-component test 9/9; browser dnd 17/17 and item-actions 20/20.
