+++ packages/chat/inventory/tree-source.js
